feat: separate transfer restrictions into namespaces
Create a general `Restrictions` namespace

BREAKING CHANGE: subdivide `restrictions` namespace into `count` and `percentage` and distributed
the methods between them
