+++ src/entities/SecurityToken/Restrictions/index.ts
@@ -1,0 +1,1 @@
+export { Restrictions } from './Restrictions';
